backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
sse-starlette==2.1.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.9

# Auth
PyJWT==2.9.0
httpx==0.27.0

# Settings
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# Redis pub/sub
redis[hiredis]==5.1.0

# Background jobs
apscheduler==3.10.4

# Discord bot
discord.py==2.4.0

# Virus scanning
pyclamd==0.4.1

# Compression
zstandard==0.23.0
